api/requirements.txt
Werkzeug-3.1.3 
blinker-1.9.0 
click-8.1.7 
flask-3.0.3 
flask-cors-5.0.0
itsdangerous-2.2.0